pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "research-etl"
version = "0.0.1"
description = "Reproducibility-focused ETL runner for research workflows"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "PyYAML>=6.0",
  "psycopg[binary]>=3.2",
  "boto3>=1.35",
  "duckdb>=1.0",
  "pyarrow>=18.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "fastapi>=0.115",
  "uvicorn>=0.30",
  "httpx>=0.27"
]
web = [
  "fastapi>=0.115",
  "uvicorn>=0.30"
]

[project.scripts]
etl = "cli:main"

[tool.setuptools]
py-modules = ["cli"]

[tool.setuptools.packages.find]
include = ["etl*"]
